feat: init models of the project
1. use dist directory as obsidian test vault
2. split into plugin, modal, settings files
3. update Makefile to manage the project
