+++ manifest.json
@@ -1,8 +1,8 @@
 {
-	"id": "obsidian-sample-plugin",
-	"name": "Sample Plugin",
+	"id": "obsidian-plugins-mng",
+	"name": "Obsidian Plugins Manager",
 	"version": "1.0.1",
-	"minAppVersion": "0.12.0",
+	"minAppVersion": "0.14.15",
 	"description": "This is a sample plugin for Obsidian. This plugin demonstrates some of the capabilities of the Obsidian API.",
 	"author": "Obsidian",
 	"authorUrl": "https://obsidian.md",
