+++ src/settings.ts
@@ -1,7 +1,9 @@
 import { App, PluginSettingTab, Setting } from "obsidian";
 
-import MyPlugin from "./main"
+import {MyPlugin} from "./plugin"
 
+// [obsidian-link-archive](https://github.com/tomzorz/obsidian-link-archive/blob/master/settings.ts)
+// [obsidian-dev-tools](https://github.com/KjellConnelly/obsidian-dev-tools)
 export class SampleSettingTab extends PluginSettingTab {
     plugin: MyPlugin;
 
@@ -11,11 +13,14 @@ export class SampleSettingTab extends PluginSettingTab {
     }
 
     display(): void {
+        const plugin: MyPlugin = this.plugin;
         const { containerEl } = this;
 
         containerEl.empty();
 
         containerEl.createEl('h2', { text: 'Settings for my awesome plugin.' });
+        containerEl.createEl("p", { text: "This plugin archives links in your note so they're available to you even if the original site goes down or gets removed." });
+        containerEl.createEl("a", { text: "Open GitHub repository", href: "https://github.com/tomzorz/obsidian-link-archive" });
 
         new Setting(containerEl)
             .setName('Setting #1')
@@ -25,8 +30,10 @@ export class SampleSettingTab extends PluginSettingTab {
                 .setValue(this.plugin.settings.mySetting)
                 .onChange(async (value) => {
                     console.log('Secret: ' + value);
-                    this.plugin.settings.mySetting = value;
+                    plugin.settings.mySetting = value;
                     await this.plugin.saveSettings();
                 }));
+
     }
+
 }
