+++ styles.css
@@ -1,4 +1,2 @@
 /* Sets all the text color to red! */
-body {
-    color: red;
-}
+body {}
